requirements.txt
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv
langchain-google-genai
langchain-core
streamlit